+++ internal/data/protogen/metrics/v1/metrics.pb.go
@@ -1723,7 +1723,7 @@ func init() {
 	proto.RegisterType((*IntSum)(nil), "opentelemetry.proto.metrics.v1.IntSum")
 	proto.RegisterType((*DoubleSum)(nil), "opentelemetry.proto.metrics.v1.DoubleSum")
 	proto.RegisterType((*IntHistogram)(nil), "opentelemetry.proto.metrics.v1.IntHistogram")
-	proto.RegisterType((*DoubleHistogram)(nil), "opentelemetry.proto.metrics.v1.DoubleHistogram")
+	proto.RegisterType((*DoubleHistogram)(nil), "opentelemetry.proto.metrics.v1.Histogram")
 	proto.RegisterType((*DoubleSummary)(nil), "opentelemetry.proto.metrics.v1.Summary")
 	proto.RegisterType((*IntDataPoint)(nil), "opentelemetry.proto.metrics.v1.IntDataPoint")
 	proto.RegisterType((*DoubleDataPoint)(nil), "opentelemetry.proto.metrics.v1.DoubleDataPoint")
@@ -4043,7 +4043,7 @@ func (m *Metric) Unmarshal(dAtA []byte) error {
 			iNdEx = postIndex
 		case 9:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field DoubleHistogram", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field Histogram", wireType)
 			}
 			var msglen int
 			for shift := uint(0); ; shift += 7 {
@@ -4672,10 +4672,10 @@ func (m *DoubleHistogram) Unmarshal(dAtA []byte) error {
 		fieldNum := int32(wire >> 3)
 		wireType := int(wire & 0x7)
 		if wireType == 4 {
-			return fmt.Errorf("proto: DoubleHistogram: wiretype end group for non-group")
+			return fmt.Errorf("proto: Histogram: wiretype end group for non-group")
 		}
 		if fieldNum <= 0 {
-			return fmt.Errorf("proto: DoubleHistogram: illegal tag %d (wire type %d)", fieldNum, wire)
+			return fmt.Errorf("proto: Histogram: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
 		case 1: